Avoid MRJAR in keycloak-core Closes #34630
